+++ metadata.rb
@@ -1,8 +1,8 @@
-name             'ubuntu-xcfe4'
+name             'ubuntu-xfce4'
 maintainer       'Lonnie VanZandt'
 maintainer_email 'lonniev@gmail.com'
 license          'Apache 2.0'
-description      'Installs/Configures Ubuntu xcfe4'
+description      'Installs/Configures Ubuntu xfce4'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
 {"ubuntu" => '12.04', "fedora" => '17.0'}.each do |os,ver|